test(chat): migrate sponsored error coverage
